requirements.txt
weasyprint==65.1
azure-storage-blob==12.25.1
beautifulsoup4==4.13.4
